refactor(theme-handler): modularize server API routes and models
The `theme-handler` server API has been refactored into a more modular and organized structure.

Previously, all theme API routes and associated response models were defined in a single large file (`src/server/api.ts`). This commit breaks down this monolithic file into a clear directory structure:

*   **`src/server/api/models/`**: Contains schema definitions for theme data (`themeModel.ts`), mutation payloads (`themePost.ts`), and standardized API responses (`themeResponses.ts`).
*   **`src/server/api/routes/`**: Contains individual files for each set of API operations: `queries.ts` (GET), `mutations.ts` (POST), `update.ts` (PUT), and `delete.ts` (DELETE).

The main `createThemeRoutes` function now composes these smaller, focused route modules, improving maintainability and readability.

Additionally, the top-level `src/index.ts` has been updated to re-export `client` and `server` modules as namespaces (`export * as client from "./client"`), providing a clearer and more explicit import structure for consumers of the package.

This refactoring aims to improve code organization, reusability of API models, and overall maintainability of the `theme-handler` package.
