+++ packages/theme-handler/src/server/api/index.ts
@@ -1,0 +1,5 @@
+export { createThemeRoutes } from "./routes"
+
+import type { createThemeRoutes } from "./routes"
+
+export type ThemeRoutesType = ReturnType<typeof createThemeRoutes>
