+++ packages/theme-handler/src/server/api/models/index.ts
@@ -1,0 +1,5 @@
+import { ThemeModel } from "./themeModel"
+import { themeMutation } from "./themePost"
+import { themeResponse } from "./themeResponses"
+
+export const theme = { model: { db: ThemeModel, post: themeMutation }, responses: themeResponse }
