+++ packages/theme-handler/src/server/api/models/themeResponses.ts
@@ -1,0 +1,26 @@
+import { t } from "elysia"
+import { theme } from "."
+
+export namespace themeResponse {
+  export const error = t.Object({
+    success: t.Literal(false),
+    error: t.String(),
+  })
+
+  export const success = {
+    list: t.Object({
+      success: t.Literal(true),
+      message: t.String(),
+      data: t.Array(theme.model),
+    }),
+    default: t.Object({
+      success: t.Literal(true),
+      message: t.String(),
+      data: theme.model,
+    }),
+    delete: t.Object({
+      success: t.Literal(true),
+      message: t.String(),
+    }),
+  }
+}
